requirements.txt
fastapi
uvicorn
GitPython
google-generativeai
python-multipart
aiofiles
chardet 
